+++ examples/hello-world/client/routes.js
@@ -1,4 +1,4 @@
-import App from './components/App'
+import App from './containers/App'
 
 export default [
   {
