+++ examples/hello-world/test/unit/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import App from '../../client/components/App'
+import App from '../../client/containers/App'
 
 test('renders correctly', () => {
   const tree = renderer.create(<App />).toJSON()
