+++ examples/todo/client/containers/App.js
@@ -1,100 +1,17 @@
 import React from 'react'
-import { connect } from 'react-redux'
-import { connectViewModel } from 'resolve-redux'
-import { bindActionCreators } from 'redux'
-import {
-  ListGroup,
-  ListGroupItem,
-  Checkbox,
-  Form,
-  Button,
-  FormControl
-} from 'react-bootstrap'
 
-import Image from './Image'
+import Header from './Header'
 
-const viewModelName = 'Todos'
-const aggregateId = 'root-id'
-
-export const App = props => {
-  const { todos, createItem, toggleItem, removeItem } = props
-
-  const placeholder = 'New Task'
-  const createItemFunc = () => {
-    createItem(aggregateId, {
-      text: newTodo.value === '' ? placeholder : newTodo.value,
-      id: Date.now()
-    })
-    newTodo.value = ''
-  }
-
-  let newTodo
-
-  return (
-    <div className="example-wrapper">
-      <h1>Tasks List</h1>
-
-      <ListGroup className="example-list">
-        {Object.keys(todos).map(id => (
-          <ListGroupItem key={id}>
-            <Checkbox
-              inline
-              checked={todos[id].checked}
-              onChange={toggleItem.bind(null, aggregateId, { id })}
-            >
-              {todos[id].text}
-            </Checkbox>
-            <Image
-              className="example-close-button"
-              src="/close-button.png"
-              onClick={removeItem.bind(null, aggregateId, { id })}
-            />
-          </ListGroupItem>
-        ))}
-      </ListGroup>
-
-      <Form inline className="example-form">
-        <FormControl
-          className="example-form-control"
-          type="text"
-          placeholder={placeholder}
-          inputRef={element => (newTodo = element)}
-          onKeyPress={event => {
-            if (event.charCode === 13) {
-              event.preventDefault()
-              createItemFunc()
-            }
-          }}
-        />
-        <Button
-          className="example-button"
-          bsStyle="success"
-          onClick={() => {
-            createItemFunc()
-          }}
-        >
-          Add Task
-        </Button>
-      </Form>
-    </div>
-  )
-}
-
-const mapStateToOptions = () => ({
-  viewModelName,
-  aggregateIds: [aggregateId]
-})
-
-const mapStateToProps = (state, { data }) => ({
-  todos: data
-})
-
-const mapDispatchToProps = (dispatch, { aggregateActions }) =>
-  bindActionCreators(aggregateActions, dispatch)
-
-export default connectViewModel(mapStateToOptions)(
-  connect(
-    mapStateToProps,
-    mapDispatchToProps
-  )(App)
+const App = ({ children }) => (
+  <div>
+    <Header
+      title="reSolve To-Do List Example"
+      name="To-Do List Example"
+      favicon="/favicon.ico"
+      css={['/bootstrap.min.css', '/style.css']}
+    />
+    {children}
+  </div>
 )
+
+export default App
