+++ examples/todo/test/unit/App.test.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { App } from '../../client/containers/App'
+import { Todos } from '../../client/containers/Todos'
 
 test('renders correctly', () => {
   const tree = renderer
     .create(
-      <App
+      <Todos
         todos={{
           id1: {
             text: 'text1',
