+++ examples/with-authentication/client/containers/App.js
@@ -1,32 +1,17 @@
 import React from 'react'
-import { connectReadModel } from 'resolve-redux'
-import { connect } from 'react-redux'
 
-import Login from '../components/Login'
-import Logout from '../components/Logout'
+import Header from './Header'
 
-export class App extends React.PureComponent {
-  render() {
-    let LoginComponent = !this.props.me.name ? Login : Logout
-
-    return (
-      <div className="example-wrapper">
-        <LoginComponent username={this.props.me && this.props.me.name} />
-      </div>
-    )
-  }
-}
-
-const mapStateToOptions = () => ({
-  readModelName: 'me',
-  resolverName: 'me',
-  resolverArgs: {}
-})
-
-const mapStateToProps = (state, { data }) => ({
-  me: data || {}
-})
-
-export default connectReadModel(mapStateToOptions)(
-  connect(mapStateToProps)(App)
+const App = ({ children }) => (
+  <div>
+    <Header
+      title="reSolve Auth Example"
+      name="Auth Example"
+      favicon="/favicon.ico"
+      css={['/bootstrap.min.css', '/style.css']}
+    />
+    {children}
+  </div>
 )
+
+export default App
