+++ examples/with-postcss/client/containers/PostCSS.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import styles from './PostCSS.css'
+
+const PostCSS = () => (
+  <div className={styles.wrapper}>
+    <div className={styles.title}>
+      Hello World, this is my first component with postcss-modules!
+    </div>
+  </div>
+)
+
+export default PostCSS
